Added Convolution code to kernel
The input image is copied tile by tile into the shared memory
of for a block of threads to efficiently access it again and again.
I have tested with 3x3 kernels so far. I believe it needs some work
to handle color images directly. But the convolution code is working.